wip(ui): move filterContext ref
wip
