+++ ui/pages/app/visualiser-les-indicateurs/par-formation/IndicateursVueFormationPage.jsx
@@ -3,9 +3,9 @@ import React from "react";
 
 import { NAVIGATION_PAGES } from "../../../../common/constants/navigationPages";
 import { BreadcrumbNav, FormationFilter, Page, Section, TerritoireFilter } from "../../../../components";
+import { useFiltersContext } from "../../../../components/_pagesComponents/FiltersContext.js";
 import FormationFilterMenu from "../../../../components/FormationFilter/FormationFilterMenu";
 import useEffectifs from "../../../../hooks/useEffectifs";
-import { useFiltersContext } from "../FiltersContext";
 import SwitchViewButton from "../SwitchViewButton";
 import IndicateursAndRepartitionFormationParCfa from "./IndicateursAndRepartitionFormationParCfa";
 import InfosFormationSection from "./InfosFormationSection";
