+++ ui/pages/app/visualiser-les-indicateurs/par-reseau/IndicateursAndRepartitionEffectifsReseau.jsx
@@ -6,6 +6,7 @@ import React from "react";
 import { fetchEffectifsDataListCsvExport } from "../../../../common/api/tableauDeBord";
 import { mapFiltersToApiFormat } from "../../../../common/utils/mapFiltersToApiFormat";
 import { RepartitionEffectifsTabs, Section } from "../../../../components";
+import { filtersPropTypes } from "../../../../components/_pagesComponents/FiltersContext.js";
 import DownloadBlock from "../../../../components/DownloadBlock/DownloadBlock";
 import RepartitionEffectifsParCfa from "../../../../components/tables/RepartitionEffectifsParCfa";
 import RepartitionEffectifsParFormation from "../../../../components/tables/RepartitionEffectifsParFormation";
@@ -13,7 +14,6 @@ import useFetchEffectifsParCfa from "../../../../hooks/useFetchEffectifsParCfa";
 import useFetchEffectifsParNiveauFormation from "../../../../hooks/useFetchEffectifsParNiveauFormation";
 import useFetchOrganismesCount from "../../../../hooks/useFetchOrganismesCount";
 import DateWithTooltipSelector from "../DateWithTooltipSelector";
-import { filtersPropTypes } from "../FiltersContext";
 import IndicateursGridStack from "../IndicateursGridStack";
 
 const IndicateursAndRepartitionEffectifsReseau = ({ filters, effectifs, loading }) => {
