+++ ui/pages/app/visualiser-les-indicateurs/par-reseau/IndicateursVueReseauPage.jsx
@@ -4,7 +4,7 @@ import React from "react";
 
 import { NAVIGATION_PAGES } from "../../../../common/constants/navigationPages";
 import { BreadcrumbNav, FormationFilter, Page, Section, TerritoireFilter } from "../../../../components";
-import { useFiltersContext } from "../FiltersContext";
+import { useFiltersContext } from "../../../../components/_pagesComponents/FiltersContext.js";
 import SwitchViewButton from "../SwitchViewButton";
 import ReseauSelect from "./ReseauSelect/ReseauSelect";
 import ReseauSelectPanel from "./ReseauSelect/ReseauSelectPanel";
