+++ ui/pages/app/visualiser-les-indicateurs/par-reseau/ReseauViewContent.jsx
@@ -1,8 +1,8 @@
 import PropTypes from "prop-types";
 import React from "react";
 
+import { useFiltersContext } from "../../../../components/_pagesComponents/FiltersContext.js";
 import useEffectifs from "../../../../hooks/useEffectifs";
-import { useFiltersContext } from "../FiltersContext";
 import IndicateursAndRepartitionEffectifsReseau from "./IndicateursAndRepartitionEffectifsReseau";
 import ReseauUpdateContactSection from "./ReseauUpdateContactSection";
